Fix for GH328 - problem with generating supporting code for inline method access
The Eclipse compiler has become more strict and we need to be informing
the code generator about variables before we use them otherwise it will
not allow them to be manipulated (even though they are there in the bytecode).

This fixes ensures sensible names are passed through from the original
point where the inline access method is created all the way to the
method binding that the generator is inspecting. These can then be used
to inform the code generator. This seems preferable to creating simply
fake entries on the backend.

Fixes #328