Fixed #327 - Attempt to push null on operand stack
This change adds tests for multiple related scenarios where the code
generation was not including correct variable information, leading
to these 'push null' problems. The new trackers for stack/variable usage
when working with recent versions of Java are more careful in that it is
not possible to manipulate stack/variable entries if they have not been
previously described to the system.  The Xwiki issue here uncovered two
but there were a number of routes through the same method that needed
testing and fixing (for example the code takes a different path for
static vs non-static and private vs non-private).

A snapshot build with these changes in seems to pass the Xwiki build
process.

Fixes #327