Fixes #323: UnresolvedType#getDimensions uses uncompiled regex, resulting in 97% of memory allocated in Pattern.compile
I say 'fixes' but all I did was use a different implementation. I don't have
a testcase that gives the memory problem but the new implementation
doesn't use Pattern so it can't give the same behaviour.

Perhaps Pattern matching is better for some situation I'm unaware of but
my microbenchmarks on this approach seem to suggest it is much faster and
the tests all seem to pass.

Fixes #323